Beautifies install script

+++ installation/ubuntu_env_setup.sh
@@ -1,33 +1,58 @@
 #!/bin/bash
 apifolder="$1"
 envname="$2"
-echo "Creating new Python environment named $envname"
+DEFAULT="$3"
+ATTENTION="$4"
+SUCCESS="$5"
+ERROR="$6"
 
-virtualenv "~/$envname"
-if [ -e "~/$envname/bin/activate" ]; then
-	. "~/$envname/bin/activate"
-	echo "Environment changed successfully! Now using Python interpreter: `which python`"
+
+if [ -d "~/$envname" ]; then
+	# if [ -e "~/$envname/"]
+	
+	virtualenv -q "$HOME/$envname"
+	echo "${SUCCESS} Created new Python environment in $HOME/$envname ${DEFAULT}"
+fi
+
+if [ -e "$HOME/$envname/bin/activate" ]; then
+	. "$HOME/$envname/bin/activate"
+	echo "${SUCCESS} Now using Python interpreter: `which python` ${DEFAULT}"
 else
-	echo "\n\n\n Something went wrong. Could not find activate script for new environment. Exitting..."
+	echo "${ERROR} Something went wrong. ${DEFAULT} Could not find activate script for new environment. Exitting... ${DEFAULT}"
 	exit 1
 fi
 
 
-
 if [ -e "$apifolder/requirements.txt" ]; then
-	echo "Installing Python modules in new environment..."
-	pip install -r "$apifolder/requirements.txt"
+	echo "${ATTENTION} Installing Python modules. This may take a while."
+	pip install -q -r "$apifolder/requirements.txt"
+	echo "${SUCCESS} Installed Python modules in $HOME/$envname/lib/python2.7/site-packages/ ${DEFAULT}"
 else
-	echo "\nCould not find requirements.txt. Exitting..."
+	echo "${ERROR} Could not find requirements.txt. Exitting..."
 	exit 1
 fi
 
 if [ -e "$apifolder/db_config_sample.yml" ]; then
-	echo "\n Copying db_config_sample.yml to db_config.yml"
-	echo "****** ATTENTION! You can db_config.yml with your own settings, but it is not necessarry."
-	cp "$apifolder/db_config_sample.yml" "$apifolder/db_config.yml"
+	if [ -e "$apifolder/db_config.yml" ]; then
+		echo "${SUCCESS} Found existing db_config.yaml. ${DEFAULT}"
+	else
+		
+		cp "$apifolder/db_config_sample.yml" "$apifolder/db_config.yml"
+		echo "${DEFAULT} Copied db_config_sample.yml to db_config.yml"
+		echo "${ATTENTION} You can edit db_config.yml with your own settings if you do not like the default settings. ${DEFAULT}"
+	fi
 else
-	echo "Could not find db_config_sample.yml. Exitting..."
+	echo "${ERROR} Could not find db_config_sample.yml. Exitting... ${DEFAULT}"
+	exit 1
+fi
+
+
+if [ -e "$apifolder/manage.py" ]; then
+	cd "$apifolder"
+	python manage.py db upgrade
+	echo "${SUCCESS} Database migration complete! ${DEFAULT}"
+else
+	echo "${ERROR} Could not find manage.py. Cannot migrate database. Exitting... ${DEFAULT}"
 	exit 1
 fi
